ui: separate MasterCheckbox and ChildCheckbox into their own files
